% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url.R
\name{api_url}
\alias{api_url}
\title{Generate an EES API URL}
\usage{
api_url(
  endpoint = "get-publications",
  publication_id = NULL,
  dataset_id = NULL,
  dataset_version = NULL,
  page_size = NULL,
  page = NULL,
  api_version = "1.0",
  environment = "dev",
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{Name of endpoint, can be "get-publications", "get-data-catalogue",
"get-summary", "get-meta", "get-data" or "query-data"}

\item{publication_id}{ID of the publication to be connected to. This is required if the
endpoint is "get-data-catalogue"}

\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "query-data"}

\item{dataset_version}{Version of data set to be connected to}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{api_version}{EES API version}

\item{environment}{EES environment to connect to: "dev" or "prod"}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
A string containing the URL for connecting to the EES API
}
\description{
Generate an EES API URL
}
\examples{
api_url()
api_url("get-publications")
api_url("get-data-catalogue", publication_id = eesyapi::example_id("publication"))
api_url("get-summary", dataset_id = eesyapi::example_id("dataset"))
api_url("get-meta", dataset_id = eesyapi::example_id("dataset"))
api_url("get-data", dataset_id = eesyapi::example_id("dataset"))
api_url("query-data", dataset_id = eesyapi::example_id("dataset"))
}
