% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_query.R
\name{api_url_query}
\alias{api_url_query}
\title{Generate URL code for API data query}
\usage{
api_url_query(
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  filter_items = NULL
)
}
\arguments{
\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{filter_items}{Filter items required as a string or vector of strings}
}
\value{
String containing data query string to append to GET data query URL
}
\description{
Generates the query code to append to the base API query URL for use by query_dataset_get().
}
\examples{
api_url_query("session_count")
api_url_query(c("session_count", "session_percent"))
}
