% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_api_filter_type.R
\name{convert_api_filter_type}
\alias{convert_api_filter_type}
\title{Convert filter type to API filter type}
\usage{
convert_api_filter_type(filter_type)
}
\arguments{
\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
String containing API friendly filter type descriptor
}
\description{
The API uses a slightly different naming convention for the different types of
filters to what is used by analysts within data files. The function just converts
from the file versions to the API versions.
}
\examples{
convert_api_filter_type("filter_items")
convert_api_filter_type("geographic_levels")
convert_api_filter_type("locations")
convert_api_filter_type("filter_items")
}
