% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_ees_environment}
\alias{default_ees_environment}
\title{Set current default EES environment}
\usage{
default_ees_environment()
}
\description{
Set current default EES environment
}
\keyword{internal}
