% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download the raw CSV for an API data set}
\usage{
download_dataset(
  dataset_id,
  dataset_version = NULL,
  ees_environment = NULL,
  api_version = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-csv", "get-data" or "post-data"}

\item{dataset_version}{Version of data set to be connected to}

\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
data.frame
}
\description{
This function is mostly designed for exploring the API, and is unlikely to
be suitable for long term production use.

There is no filtering down of the file so you will always get the whole file
and in some instances this may be very large.

As there are no IDs involved, this is brittle and code relying on this
function will likely break whenever there is renaming of variables or items
in the data.

It is recommended to take the time to set up custom queries using the
\code{query_dataset()} function instead. If you are using this function for more
than exploratory purposes, make sure you subscribe to the data set you're
downloading and then keep track of any updates to the data.
}
\details{
This gives a super quick way to just fetch the whole file in a human
readable format.
}
\examples{
download_dataset(example_id("dataset"))
}
