% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesapi_url.R
\name{eesapi_url_pages}
\alias{eesapi_url_pages}
\title{EES API URL paging}
\usage{
eesapi_url_pages(page_size = 40, page = NULL)
}
\arguments{
\item{page_size}{Number of results to return in a single query (max 40)}

\item{page}{Page number to return}
}
\value{
String containing pages query
}
\description{
EES API URL paging
}
\examples{
eesapi_url_pages()
eesapi_url_pages(page_size = 20, page = 2)
}
