% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesmeta_to_filtertable.R
\name{eesmeta_to_filtertable}
\alias{eesmeta_to_filtertable}
\title{EES meta to filter table}
\usage{
eesmeta_to_filtertable(api_meta)
}
\arguments{
\item{api_meta}{}
}
\value{
Data frame containing the data set filters
}
\description{
EES meta to filter table
}
