% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesquery_to_dataframe.R
\name{eesquery_to_dataframe}
\alias{eesquery_to_dataframe}
\title{EES query output to data.frame}
\usage{
eesquery_to_dataframe(api_output, clear_codes = TRUE)
}
\arguments{
\item{api_output}{The results of an EES API data query}

\item{clear_codes}{Flag as to whether to clear the API code tags from the
output}
}
\value{
data.frame containing the API query results
}
\description{
EES query output to data.frame
}
