% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesyapi-package.R
\docType{package}
\name{eesyapi-package}
\alias{eesyapi}
\alias{eesyapi-package}
\title{eesyapi: EES-y API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package with useful utility functions for connecting to, and processing data from, the DfE's explore education statistics API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfe-analytical-services.github.io/eesyapi}
}

}
\author{
\strong{Maintainer}: Rich Bielby \email{richard.bielby@education.gov.uk} (\href{https://orcid.org/0000-0001-9070-9969}{ORCID})

Authors:
\itemize{
  \item Cam Race \email{cameron.race@education.gov.uk}
}

Other contributors:
\itemize{
  \item Laura Selby \email{laura.selby@education.gov.uk} [contributor]
}

}
\keyword{internal}
