% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_data_raw}
\alias{example_data_raw}
\title{Title}
\usage{
example_data_raw(group = "attendance", size = 32)
}
\arguments{
\item{group}{Choose the publication group of examples to use. Can be "attendance".}

\item{size}{Number of rows to return (max = 1000)}
}
\value{
Nested list form of example data from the API
}
\description{
Title
}
\examples{
example_data_raw()
}
