% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_environment}
\alias{example_environment}
\title{Example environment}
\usage{
example_environment()
}
\value{
String containing default environment
}
\description{
Example environment
}
\examples{
example_environment()
}
