% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_geography_query}
\alias{example_geography_query}
\title{Create an example geography-query data frame}
\usage{
example_geography_query(level = "nat_yorks")
}
\arguments{
\item{level}{Query level within available options, can be one of \"nat_yorks\" or
\"nat_yorks_yorkslas\"}
}
\value{
Data frame containing an example geography query
}
\description{
Create an example geography-query data frame
}
\examples{
example_geography_query()
}
