% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_catalogue.R
\name{get_data_catalogue}
\alias{get_data_catalogue}
\title{Get data set catalogue}
\usage{
get_data_catalogue(publication_id = NULL)
}
\arguments{
\item{publication_id}{The publication ID, leave as NULL to query across all publications}
}
\value{
Data frame listing data sets
}
\description{
Get data set catalogue
}
\examples{
get_data_catalogue()
}
