% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Get publication specific data set catalogue}
\usage{
get_datasets(publication_id, page_size = NULL, page = NULL, verbose = FALSE)
}
\arguments{
\item{publication_id}{The publication ID as used by the API}

\item{page_size}{Number of results to collect in a single query to the API (max 40)}

\item{page}{Page number to return (Default is NULL which will loop through until all
pages of the query are collated).}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame listing the data sets contained within a single publication
}
\description{
Get publication specific data set catalogue
}
\examples{
get_datasets(example_id("publication"))
}
