% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{get_meta}
\usage{
get_meta(dataset_id, dataset_version = NULL, api_version = NULL, parse = TRUE)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{parse}{Parse result into structured list}
}
\value{
Results of query to meta data
}
\description{
get_meta
}
\examples{
get_meta("d7329101-f275-d277-bbfe-d8cfaa709833")
}
