% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta_response}
\alias{get_meta_response}
\title{Get the base API response for a data set's meta data}
\usage{
get_meta_response(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  parse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "post-data"}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{parse}{Parse result into structured list}

\item{verbose}{Run with additional contextual messaging, logical, default = FALSE}
}
\value{
Results of query to API meta data endpoint
}
\description{
Get the metadata information for a data set available from the EES API.
}
\examples{
get_meta_response(example_id())
}
