% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication.R
\name{get_publication_catalogue}
\alias{get_publication_catalogue}
\title{Get publications}
\usage{
get_publication_catalogue(page_size = 40, page = NULL, verbose = FALSE)
}
\arguments{
\item{page_size}{Number of results to return in a single query (max 40)}

\item{page}{Page number to return}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame listing all available publications
}
\description{
Get publications
}
\examples{
get_publication_catalogue()
}
