% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_catalogue.R
\name{get_publication_dataset_list}
\alias{get_publication_dataset_list}
\title{Get publication specific data set catalogue}
\usage{
get_publication_dataset_list(publication_id)
}
\arguments{
\item{publication_id}{The publication ID as used by the API}
}
\value{
Data frame listing the data sets contained within a single publication
}
\description{
Get publication specific data set catalogue
}
\examples{
get_publication_dataset_list(example_id("publication"))
}
