% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_catalogue.R
\name{get_publication_datasets}
\alias{get_publication_datasets}
\title{Get publication specific data set catalogue}
\usage{
get_publication_datasets(
  publication_id,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{publication_id}{The publication ID as used by the API}

\item{page_size}{Number of results to return in a single query (max 40)}

\item{page}{Page number to return}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame listing the data sets contained within a single publication
}
\description{
Get publication specific data set catalogue
}
\examples{
get_publication_datasets(example_id("publication"))
}
