% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications}
\alias{get_publications}
\title{Get publications}
\usage{
get_publications(
  ees_environment = NULL,
  api_version = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
Data frame listing all available publications
}
\description{
Get publications
}
\examples{
get_publications()
}
