% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/http_request_error.R
\name{http_request_error}
\alias{http_request_error}
\title{http request error}
\usage{
http_request_error(response_status)

http_request_error(response_status)
}
\arguments{
\item{response_status}{Response code from an API request}
}
\value{
Translation of the response code

Translation of the response code
}
\description{
Translate a http error code into an error message.

Translate a http error code into an error message.
}
\examples{
http_request_error(200)
http_request_error(200)
http_request_error(400)
http_request_error(504)
}
