% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_api_dataset.R
\name{parse_api_dataset}
\alias{parse_api_dataset}
\title{Title}
\usage{
parse_api_dataset(api_data_result, dataset_id = NULL)
}
\arguments{
\item{api_data_result}{A json data result list as returned from the API}

\item{dataset_id}{ID of data set to be connected to.}
}
\value{
Data frame containing API data results
}
\description{
This takes the json results output from the API and converts it into a usable data frame.
It's primary use is as a helper to get_dataset and post_dataset, but can be used in
isolation from those by running an API query and passing the following to
parse_api_dataset():
response |>
httr::content("text") |>
jsonlite::fromJSON() |>
parse_api_dataset()
}
\examples{
get_dataset(example_id(), indicators = example_id("indicator"), parse = FALSE) |>
  parse_api_dataset()
}
