% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data_squids.R
\name{parse_data_squids}
\alias{parse_data_squids}
\title{Parse IDs in returned API data}
\usage{
parse_data_squids(data, datatset_id)
}
\arguments{
\item{data}{Data frame containing data as returned from the API by \code{get_dataset()} or
\code{post_dataset()}}

\item{datatset_id}{String containing the data set ID}
}
\value{
Data frame
}
\description{
The API uses unique IDs (squids) to identify each filter column, filter_item and indicator column.
This facilitates continuity, i.e. data creators may need to change the col_name and labeling in
their data files, whilst maintaining the same fundamental content. Results from the API are given
using these IDs, but it's expected that analysts will want to translate these back into the
col_names assigned by analysts. \code{parse_data_squids()} takes a data frame extracted from the API
and translates all the squids in that data frame to their assigned col_names based on the meta
data available from the API.
}
\examples{
get_dataset(example_id(), indicators = example_id("indicator"), page = 1) |>
  parse_data_squids(example_id())
}
