% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_query_utils.R
\name{parse_filter_in}
\alias{parse_filter_in}
\title{Create \if{html}{\out{<filter>}}.in query string for URL queries}
\usage{
parse_filter_in(items = NULL, type = "filters")
}
\arguments{
\item{items}{items to be included in the "in" statement}

\item{type}{type of filter items being queried: "time_periods", "geographic_levels",
"locations" or "filters"}
}
\value{
Query string for use in URL based API queries
}
\description{
Outputs a URL query string containing timePeriods.in=..., geographicLevels.in=..., etc for
use with querying a data set via GET.
}
\examples{
parse_filter_in(c("2024|W11", "2024|W12"), type = "time_period")
}
