% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_filter_to_json}
\alias{parse_filter_to_json}
\title{Parse geographic levels to json}
\usage{
parse_filter_to_json(filter_items, filter_type = "filter_items")
}
\arguments{
\item{filter_items}{Filter items required as a string or vector of strings}
}
\value{
String containing json form query for geographic levels
}
\description{
Create a json query sub-string based on geographic levels constraints
}
\examples{
parse_filter_to_json(c("NAT", "REG"), filter_type = "geographic_levels")
}
