% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_geographic_levels_to_json}
\alias{parse_geographic_levels_to_json}
\title{Parse geographic levels to json}
\usage{
parse_geographic_levels_to_json(geographic_levels)
}
\arguments{
\item{geographic_levels}{Geographic levels required as a string or vector of strings}
}
\value{
String containing json form query for geographic levels
}
\description{
Create a json query sub-string based on geographic levels constraints
}
\examples{
parse_geographic_levels_to_json(c("NAT", "REG"))
}
