% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_geographies_to_json}
\alias{parse_geographies_to_json}
\title{Parse geographies to json}
\usage{
parse_geographies_to_json(geographies)
}
\value{
String containing json form query for geographies
}
\description{
Create a json query sub-string based on location constraints
}
\examples{
parse_geographies_to_json(c("NAT|id|dP0Zw", "REG|id|rg3Nj")) |>
  cat()
}
