% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_locations_to_json}
\alias{parse_locations_to_json}
\title{Parse locations to json}
\usage{
parse_locations_to_json(locations)
}
\arguments{
\item{locations}{Location code required as a string or vector of strings}
}
\value{
String containing json form query for locations
}
\description{
Create a json query sub-string based on location constraints
}
\examples{
parse_locations_to_json(c("NAT|id|dP0Zw", "REG|id|rg3Nj")) |>
  cat()
}
