% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_filter_columns}
\alias{parse_meta_filter_columns}
\title{Parse API meta to give the filter columns}
\usage{
parse_meta_filter_columns(api_meta_filters)
}
\arguments{
\item{api_meta_filters}{Filter information provided by the API output}
}
\value{
data frame containing column names and labels
}
\description{
Parse API meta to give the filter columns
}
\examples{
get_meta_response("d7329101-f275-d277-bbfe-d8cfaa709833")$filters |>
  parse_meta_filter_columns()
}
