% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_filter_item_codes}
\alias{parse_meta_filter_item_codes}
\title{Parse API meta to give the filter item codes}
\usage{
parse_meta_filter_item_codes(api_meta_filters)
}
\arguments{
\item{api_meta_filters}{Filter information provided by the API output}
}
\value{
Data frame containing filter item codes matched to filter item labels and col_name
}
\description{
Parse API meta to give the filter item codes
}
\examples{
get_meta_response("d7329101-f275-d277-bbfe-d8cfaa709833")$filters |>
  parse_meta_filter_item_codes()
}
