% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_filter_item_ids}
\alias{parse_meta_filter_item_ids}
\title{Parse API meta to give the filter item codes}
\usage{
parse_meta_filter_item_ids(api_meta_filters, verbose = FALSE)
}
\arguments{
\item{api_meta_filters}{Filter information provided by the API output}

\item{verbose}{Run with additional contextual messaging, logical, default = FALSE#'}
}
\value{
Data frame containing filter item codes matched to filter item labels and col_name
}
\description{
Parse API meta to give the filter item codes
}
\examples{
get_meta_response(example_id())$filters |>
  parse_meta_filter_item_ids()
}
