% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_location_ids}
\alias{parse_meta_location_ids}
\title{Parse API meta to give the locations}
\usage{
parse_meta_location_ids(api_meta_locations)
}
\arguments{
\item{api_meta_locations}{Locations information provided by the API output}
}
\value{
Data frame containing location item codes matched
}
\description{
Parse API meta to give the locations
}
\examples{
get_meta_response("d7329101-f275-d277-bbfe-d8cfaa709833")$locations |>
  parse_meta_location_ids()
}
