% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_params_to_json}
\alias{parse_params_to_json}
\title{Create a json query string}
\usage{
parse_params_to_json(
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL
)
}
\value{
String containing json query body for use with http POST request
}
\description{
A short description...
}
\examples{
parse_params_to_json(example_id("indicator"))
}
