% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_sqids_dataset}
\alias{parse_sqids_dataset}
\title{Parse IDs in returned API data}
\usage{
parse_sqids_dataset(data, dataset_id, verbose = FALSE)
}
\arguments{
\item{data}{Data frame containing data as returned from the API by \code{get_dataset()} or
\code{post_dataset()}}

\item{dataset_id}{String containing the data set ID}

\item{verbose}{Output status messaging for user}
}
\value{
Data frame
}
\description{
The API uses unique IDs (sqids) to identify each filter column, filter_item and indicator
column. This facilitates continuity, i.e. data creators may need to change the col_name and
labeling in their data files, whilst maintaining the same fundamental content. Results from the
API are given using these IDs, but it's expected that analysts will want to translate these back
into the col_names assigned by analysts. \code{parse_data_sqids()} takes a data frame extracted from
the API and translates all the sqids in that data frame to their assigned col_names based on
the meta data available from the API.
}
\examples{
get_dataset(example_id(), indicators = example_id("indicator"), page = 1) |>
  parse_sqids_dataset(example_id())
}
