% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_sqids_geographies}
\alias{parse_sqids_geographies}
\title{Parse geography sqids}
\usage{
parse_sqids_geographies(geographies, meta, verbose = FALSE)
}
\arguments{
\item{geographies}{}

\item{meta}{Meta data for the data set as provided by \code{get_meta()}}

\item{verbose}{Run in verbose mode with debugging messages}
}
\value{
Data frame of parsed geography information
}
\description{
Parse geography sqids
}
\examples{
example_data_raw() |>
  magrittr::use_series("locations") |>
  parse_sqids_geographies(get_meta(example_id(group = "attendance")))
}
