% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_sqids_indicator}
\alias{parse_sqids_indicator}
\title{Parse IDs in a filter column}
\usage{
parse_sqids_indicator(column_sqid, meta, verbose = FALSE)
}
\arguments{
\item{column_sqid}{The indicator col_id}

\item{meta}{Meta data for the data set as provided by \code{get_meta()}}

\item{verbose}{Output status messaging for user}
}
\value{
Data frame
}
\description{
The API uses unique IDs (sqids) to identify each filter column and its contents (filter items).
This function parses those into the data creators id and item labels based on the meta data
stored on the API for the data set.
}
\examples{
parse_sqids_indicator(
  example_id("indicator"),
  get_meta(example_id()),
)
}
