% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_sqids_indicators}
\alias{parse_sqids_indicators}
\title{Parse IDs in a set of indicators}
\usage{
parse_sqids_indicators(indicators, meta, verbose = FALSE)
}
\arguments{
\item{indicators}{A set of indicator columns as taken from a data set downloaded from the API}

\item{meta}{Meta data for the data set as provided by \code{get_meta()}}

\item{verbose}{Run in verbose mode with debugging messages}
}
\value{
Data frame
}
\description{
The API uses unique IDs (sqids) to identify each indicator column. This function parses those
into the data creators' column names based on the meta data stored on the API for the data set.
}
\examples{
example_data_raw(group = "attendance") |>
  magrittr::use_series("values") |>
  parse_sqids_indicators(get_meta(example_id(group = "attendance")))
}
