% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sqids.R
\name{parse_sqids_locations}
\alias{parse_sqids_locations}
\title{Parse location sqids}
\usage{
parse_sqids_locations(locations, meta, verbose = FALSE)
}
\arguments{
\item{locations}{A set of location columns as taken from a data set downloaded from the API}

\item{meta}{Meta data for the data set as provided by \code{get_meta()}}

\item{verbose}{Run in verbose mode with debugging messages}
}
\value{
Data frame of parsed geography information
}
\description{
The API uses unique IDs (sqids) to identify each location in a data set. This function parses
those into the corresponding location codes and names based on the meta data stored on the API
for the data set.
}
\examples{
example_data_raw() |>
  magrittr::use_series("locations") |>
  parse_sqids_locations(get_meta(example_id(group = "attendance")))
}
