% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_filter_eq}
\alias{parse_tojson_filter_eq}
\title{Parse geographic levels to json}
\usage{
parse_tojson_filter_eq(filter_item, filter_type = "filter_item")
}
\arguments{
\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
String containing json form query for geographic levels
}
\description{
Create a json query sub-string based on geographic levels constraints
}
\examples{
parse_tojson_filter_eq("NAT", filter_type = "geographic_levels") |> cat()
}
