% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_filter_in}
\alias{parse_tojson_filter_in}
\title{Parse a filter-in type query to json}
\usage{
parse_tojson_filter_in(items, filter_type = "filter_items")
}
\arguments{
\item{items}{items to be included in the "in" statement}

\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
String containing json form query based on filter-in constraints
}
\description{
Create a json query sub-string based on filter-in constraints
}
\examples{
parse_tojson_filter_in(c("NAT", "REG"), filter_type = "geographic_levels")
}
