% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_filter_list}
\alias{parse_tojson_filter_list}
\title{Parse a combination-filter query to json}
\usage{
parse_tojson_filter_list(items, filter_type = "filter_items")
}
\arguments{
\item{items}{items to be included in the "in" statement}

\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
String containing json form query with \"and\"-combination of different filter
selections
}
\description{
Create a json query sub-string based on a combination \"in\" and \"and\" constraints
}
\examples{
parse_tojson_filter_list(
  list(
    day_number = c("uLQo4", "qf0jG", "aMjLP"),
    reason = c("bBrtT", "ThjPJ", "hsHyW", "m2m9K"),
    education_phase = c("5UNdi", "crH31")
  )
) |>
  cat()
}
