% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_indicators}
\alias{parse_tojson_indicators}
\title{Parse an indicator-in type query to json}
\usage{
parse_tojson_indicators(indicators)
}
\arguments{
\item{indicators}{String or vector of strings containing indicator ids}
}
\value{
A json query string to select a set of indicators
}
\description{
Create a json query sub-string based on indicator-in constraints
}
\examples{
parse_tojson_indicators(example_id("indicator")) |>
  cat()
}
