% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_location}
\alias{parse_tojson_location}
\title{Create json location search string from geographies}
\usage{
parse_tojson_location(geographies, include_comma = FALSE)
}
\arguments{
\item{geographies}{Vector or data frame of search geographies}

\item{include_comma}{Include a comma before return strings (logical)}
}
\value{
Vector of strings containing json location search string
}
\description{
Create json location search string from geographies
}
\examples{
parse_tojson_location(example_geography_query()) |> cat()
}
