% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_dataset.R
\name{preview_dataset}
\alias{preview_dataset}
\title{Preview the raw CSV for an API data set}
\usage{
preview_dataset(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  n_max = 10,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set}

\item{dataset_version}{Version number of data set}

\item{api_version}{EES API version}

\item{n_max}{maximum number of rows to preview, 10 by default, Inf will get
all available rows}

\item{verbose}{Run with additional contextual messaging, logical,
default = FALSE}
}
\value{
data.frame
}
\description{
This function is mostly designed for exploring the API, and is unlikely to
be suitable for long term production use.

You can set the number of rows to preview using the n_max parameter. This
uses the n_max from \code{readr::read_csv()} under the hood.

As there are no IDs involved, this is brittle and code relying on this
function will likely break whenever there is renaming of variables or items
in the data.

It is recommended to take the time to set up custom queries using the
\code{query_dataset()} function instead.

If you are using this function for more than exploratory purposes, make
sure you subscribe to the data set you're downloading and then keep track
of any updates to the data.
}
\details{
This gives a super quick way to just fetch the file in a human readable
format.
}
\examples{
# Preview first 10 rows
preview_dataset(example_id("dataset"))

# Get 2 rows
preview_dataset(example_id("dataset"), n_max = 2)

# Get all rows
preview_dataset(example_id("dataset"), n_max = Inf)
}
