% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_get_dataset.R
\name{query_get_dataset}
\alias{query_get_dataset}
\title{Query a data set using GET and a query URL}
\usage{
query_get_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  dataset_version = NULL,
  api_version = NULL
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "query-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}
}
\value{
Data frame containing query results of an API data set
}
\description{
This function provides a method for generating and sending a URL based data query to the
EES API. As a minimum, it requires the dataset_id and indicators flags to be provided.
}
\examples{
query_get_dataset(example_id(), indicators = "session_count")
}
