% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_api_version}
\alias{validate_api_version}
\title{Validate API version}
\usage{
validate_api_version(api_version)
}
\arguments{
\item{api_version}{Variable containing the api_version}
}
\description{
Validate API version
}
\examples{
validate_api_version(1.0)
}
