% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_ees_environment}
\alias{validate_ees_environment}
\title{Validate EES environment}
\usage{
validate_ees_environment(ees_environment)
}
\arguments{
\item{ees_environment}{String containing the ees_environment name: "dev", "test", "preprod" or
"prod"}
}
\description{
Validate EES environment
}
\examples{
validate_ees_environment("prod")
}
