% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_ees_filter_type}
\alias{validate_ees_filter_type}
\title{Validate filter type}
\usage{
validate_ees_filter_type(filter_type)
}
\arguments{
\item{filter_type}{}
}
\description{
Validate filter type
}
\examples{
validate_ees_filter_type("time_periods")
}
