% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_ees_id}
\alias{validate_ees_id}
\title{Validate element IDs}
\usage{
validate_ees_id(element_id, level = "publication")
}
\arguments{
\item{element_id}{ID for publication or a data set}

\item{level}{ID level: "publication" or "dataset"}
}
\description{
Validate element IDs
}
\examples{
validate_ees_id(example_id("publication"), level = "publication")
}
