% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{validate_page_size}
\alias{validate_page_size}
\title{Validate page size}
\usage{
validate_page_size(page_size, min = 1, max = 40)
}
\arguments{
\item{page_size}{Chosen page size}

\item{min}{Minimum valid page_size for EES API}

\item{max}{Maximum valid page_size for EES API}
}
\value{
Logic
}
\description{
Validate page size
}
\examples{
validate_page_size(20)
}
