% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_max_pages.R
\name{warning_max_pages}
\alias{warning_max_pages}
\title{Overshot maximum pages in results}
\usage{
warning_max_pages(api_result)
}
\arguments{
\item{api_result}{Output from an API get query}
}
\description{
Overshot maximum pages in results
}
\examples{
response_page <- httr::GET(eesyapi::api_url(page_size = 10, page = 1)) |>
  httr::content("text") |>
  jsonlite::fromJSON() |>
  warning_max_pages()
}
