% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url.R
\name{api_url}
\alias{api_url}
\title{Generate an EES API URL}
\usage{
api_url(
  endpoint = "get-publications",
  publication_id = NULL,
  dataset_id = NULL,
  indicators = NULL,
  time_periods = NULL,
  geographic_levels = NULL,
  dataset_version = NULL,
  page_size = NULL,
  page = NULL,
  api_version = "1.0",
  environment = "dev",
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{Name of endpoint, can be "get-publications", "get-data-catalogue",
"get-summary", "get-meta", "get-data" or "query-data"}

\item{publication_id}{ID of the publication to be connected to. This is required if the
endpoint is "get-data-catalogue"}

\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "query-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{dataset_version}{Version of data set to be connected to}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{api_version}{EES API version}

\item{environment}{EES environment to connect to: "dev", "test", "preprod" or "prod"}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
A string containing the URL for connecting to the EES API
}
\description{
This function returns a single URL to connect to a chosen EES API endpoint. The resulting URL can
be used with GET or POST (as appropriate) in order to connect. Whether the publication_id or
dataset_id (or neither) parameter are required depends on the endpoint chose.

\tabular{lr}{
\strong{Endpoints} \tab \strong{id required} \cr
get-publications \tab Neither  \cr
get-data-catalogue \tab publication_id  \cr
get-summary, get-meta, get-data, query-data \tab dataset_id  \cr
}
}
\examples{
api_url()
api_url("get-publications")
api_url("get-data-catalogue", publication_id = eesyapi::example_id("publication"))
api_url("get-summary", dataset_id = eesyapi::example_id("dataset"))
api_url("get-meta", dataset_id = eesyapi::example_id("dataset"))
api_url("get-data", dataset_id = eesyapi::example_id("dataset"))
api_url("query-data", dataset_id = eesyapi::example_id("dataset"))
}
