% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download the raw CSV for an API data set}
\usage{
download_dataset(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set}

\item{dataset_version}{Version number of data set}

\item{api_version}{EES API version}

\item{verbose}{Run with additional contextual messaging, logical, default = FALSE}
}
\value{
data.frame
}
\description{
This function is mostly designed for exploring the API, and is unlikely to
be suitable for long term production use.

There is no filtering down of the file so you will always get the whole file
and in some instances this may be very large.

As there are no IDs involved, this is brittle and code relying on this
function will likely break whenever there is renaming of variables or items
in the data.

It is recommended to subscribe to the data set you're querying and then keep
track of any updates, including minor updates to the data.
}
\details{
This gives a super quick way to just fetch the whole file in a human
readable format.
}
\examples{
download_dataset(example_id("dataset", group = "public-api-testing"))
}
