% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesapi_url.R
\name{eesapi_url}
\alias{eesapi_url}
\title{eesapi_url}
\usage{
eesapi_url(api_version = "1.0")
}
\arguments{
\item{api_version}{EES API version}
}
\value{
The EES URL for connecting to the API
}
\description{
eesapi_url
}
\examples{
eesapi_url()
}
