% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_data_raw}
\alias{example_data_raw}
\title{Example raw data}
\usage{
example_data_raw(group = "attendance", size = 32)
}
\arguments{
\item{group}{Choose the publication group of examples to use. Can be "attendance".}

\item{size}{Number of rows to return (max = 1000)}
}
\value{
Nested list form of example data from the API
}
\description{
Download some example raw data. Mainly intended for use in developing / testing the sqid parsing
or as an example of getting raw data if any end users would prefer to do the sqid parsing
themselves.
}
\examples{
example_data_raw()
}
