% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_id}
\alias{example_id}
\title{Example ID}
\usage{
example_id(
  level = "dataset",
  environment = "dev",
  group = "public-api-testing"
)
}
\arguments{
\item{level}{Level of ID example to return. A range are available, although not every example
group necessarily contains all possible examples. The following are generally available.
\itemize{
\item "all": Return full list of example options for given group.
\item "publication": Return example publication ID
\item "dataset": Return example data set ID
\item "time_period" / "time_periods": Return example time_period(s)
\item "location_id" / "location_ids": Return example location ID(s)
\item "location_code / location_codes": Return example location code(s)
\item "filter": Return example filter column ID
\item "filter_item" / "filter_items_short" / "filter_items_long": Return example filter ID or
example short / long filter query list.
\item "indicator": Return example indicator ID
}}

\item{environment}{Environment to return a working example for: "dev" or "test"}

\item{group}{Choose the publication group of examples to use. Options are:
\itemize{
\item "attendance": Large example data set, careful what you ask for
\item "public-api-testing": Smaller example data set
}}
}
\value{
String, vector or list containing example ID(s) present in the API
}
\description{
This function returns examples of working IDs that can be used with the eesyapi functions, such
as query_dataset().
}
\examples{
example_id("all")
example_id()
example_id("publication")
example_id("publication", group = "attendance")
example_id("time_period", group = "attendance")
example_id("location_ids", group = "attendance")
example_id("filter_items_short", group = "attendance")
example_id("indicator", group = "attendance")
}
