% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_id.R
\name{example_id}
\alias{example_id}
\title{Example ID}
\usage{
example_id(level = "dataset", environment = "dev", group = "attendance")
}
\arguments{
\item{level}{Level of ID example to return: "publication" or "data set"}

\item{environment}{Environment to return a working example for}

\item{group}{Choose the publication group of examples to use. Can be "attendance".}
}
\value{
String containing an example ID present in the API
}
\description{
This function returns examples of working IDs that can be used with the API.
}
\examples{
example_id()
}
