% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_catalogue.R
\name{get_data_catalogue}
\alias{get_data_catalogue}
\title{Get publication specific data set catalogue}
\usage{
get_data_catalogue(
  publication_id,
  ees_environment = NULL,
  api_version = NULL,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{publication_id}{ID of the publication to be connected to. This is required if the
endpoint is "get-data-catalogue"}

\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
Data frame listing the data sets contained within a single publication
}
\description{
Get publication specific data set catalogue
}
\examples{
get_data_catalogue(example_id("publication"))
}
