% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Query a data set using GET and a query URL}
\usage{
get_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  dataset_version = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 1000,
  parse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "post-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{locations}{Location code required as a string or vector of strings}

\item{filter_items}{Filter items required as a string or vector of strings}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{page}{Page number of query results to return}

\item{page_size}{Number of results to return in a single query}

\item{parse}{Logical flag to activate parsing of the results. Default: TRUE}

\item{verbose}{Run with additional contextual messaging, logical, default = FALSE#'}
}
\value{
Data frame containing query results of an API data set
}
\description{
This function provides a method for generating and sending a URL based data query to the
EES API. As a minimum, it requires the dataset_id and indicators flags to be provided.

Note that the GET command is very limited on the level of logic it can process. For example
there is no way of using GET to make a query that combines different filters with AND logic.
So if you give GET a set of filter items to search on, it can only return all rows containing
any of those items.
}
\examples{
get_dataset(
  example_id(),
  geographic_levels = "NAT",
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator")
)
}
