% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Title}
\usage{
get_meta(dataset_id, dataset_version = NULL, api_version = NULL)
}
\arguments{
\item{dataset_id}{ID of dataset to be connected to}

\item{dataset_version}{Verson of dataset to be connected to}

\item{api_version}{EES API version}
}
\value{
Results of query to meta data
}
\description{
Title
}
