% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta_response}
\alias{get_meta_response}
\title{Get the base API response for a data set's meta data}
\usage{
get_meta_response(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  parse = TRUE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{parse}{Parse result into structured list}
}
\value{
Results of query to API meta data endpoint
}
\description{
Get the metadata information for a data set available from the EES API.
}
\examples{
get_meta_response("d7329101-f275-d277-bbfe-d8cfaa709833")
}
