% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications}
\alias{get_publications}
\title{Get publications}
\usage{
get_publications(page_size = 40, page = NULL, verbose = FALSE)
}
\arguments{
\item{page_size}{Number of results to collect in a single query to the API (max 40)}

\item{page}{Page number to return (Default is NULL which will loop through until all
pages of the query are collated).}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame listing all available publications
}
\description{
Get publications
}
\examples{
get_publications()
}
