% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_catalogue.R
\name{get_publications}
\alias{get_publications}
\title{Get publications}
\usage{
get_publications(page_size = 40)
}
\arguments{
\item{page_size}{Number of results to return in a single query}
}
\value{
Data frame listing all available publications
}
\description{
Get publications
}
\examples{
get_publications()
}
