% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_request_error.R
\name{http_request_error}
\alias{http_request_error}
\title{Contextualise http request errors}
\usage{
http_request_error(response, verbose = FALSE)
}
\arguments{
\item{response}{HTTP response from the API}

\item{verbose}{Run in verbose mode, logical, default = FALSE}
}
\value{
Translation of the response code
}
\description{
Translate a http error code into an error message.
}
\examples{
http_request_error(200)
http_request_error(400)
http_request_error(504)
}
