% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_query_utils.R
\name{parse_filter_in}
\alias{parse_filter_in}
\title{Create "filter in" query string for url queries}
\usage{
parse_filter_in(items = NULL, type = "filters")
}
\arguments{
\item{items}{items to be included in the "in" statement}

\item{type}{type of filter items being queried: "time_periods", "geographic_levels", "locations" or
"filters"}
}
\value{
Query string for use in URL based API queries
}
\description{
A short description...
}
\examples{
parse_standard_filter_in(c("2024|W11", "2024|W12"), type = "time_period")
}
