% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_query_utils.R
\name{parse_filter_in}
\alias{parse_filter_in}
\title{Create \if{html}{\out{<filter>}}.in query string for URL queries}
\usage{
parse_filter_in(items, filter_type)
}
\arguments{
\item{items}{items to be included in the "in" statement}

\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
Query string for use in URL based API queries
}
\description{
Outputs a URL query string containing timePeriods.in=..., geographicLevels.in=..., etc for
use with querying a data set via GET.
}
\examples{
parse_filter_in(c("2024|W11", "2024|W12"), filter_type = "time_periods")
}
