% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_indicator_columns}
\alias{parse_meta_indicator_columns}
\title{Parse API meta to give the indicator columns}
\usage{
parse_meta_indicator_columns(api_meta_indicators)
}
\arguments{
\item{api_meta_indicators}{Filter information provided by the API output}
}
\value{
data frame containing indicator column names and labels
}
\description{
Parse API meta to give the indicator columns
}
\examples{
get_meta_response("d7329101-f275-d277-bbfe-d8cfaa709833")$indicators |>
  parse_meta_indicator_columns()
}
