% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{parse_meta_time_periods}
\alias{parse_meta_time_periods}
\title{Parse API meta to give the time periods}
\usage{
parse_meta_time_periods(api_meta_time_periods, verbose = FALSE)
}
\arguments{
\item{api_meta_time_periods}{Time periods information provided by the API output}

\item{verbose}{Run with additional contextual messaging, logical, default = FALSE}
}
\value{
Data frame containing location item codes matched
}
\description{
Parse API meta to give the time periods
}
\examples{
get_meta_response(example_id())$timePeriods |>
  parse_meta_time_periods()
}
