% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_params_to_json}
\alias{parse_params_to_json}
\title{Create a json query string}
\usage{
parse_params_to_json(
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  page = 1,
  page_size = 1000
)
}
\arguments{
\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{locations}{Location code required as a string or vector of strings}

\item{filter_items}{Filter items required as a string or vector of strings}

\item{page}{Page number of query results to return}

\item{page_size}{Number of results to return in a single query}
}
\value{
String containing json query body for use with http POST request
}
\description{
A short description...
}
\examples{
parse_params_to_json(example_id("indicator"))
}
