% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_time_periods_to_json}
\alias{parse_time_periods_to_json}
\title{Parse time_periods to json}
\usage{
parse_time_periods_to_json(time_periods)
}
\arguments{
\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}
}
\value{
String containing json form query for time periods
}
\description{
Create a json query sub-string based on time periods constraints
}
\examples{
parse_time_periods_to_json(c("2023|W25", "2024|W12"))
}
