% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset_utils.R
\name{parse_tojson_filter_in}
\alias{parse_tojson_filter_in}
\title{Parse geographic levels to json}
\usage{
parse_tojson_filter_in(filter_items, filter_type = "filter_items")
}
\arguments{
\item{filter_items}{Filter items required as a string or vector of strings}

\item{filter_type}{type of filter being queried: "time_periods", "geographic_levels",
"locations" or "filter_items"}
}
\value{
String containing json form query for geographic levels
}
\description{
Create a json query sub-string based on geographic levels constraints
}
\examples{
parse_tojson_filter_in(c("NAT", "REG"), filter_type = "geographic_levels")
}
