% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset.R
\name{post_dataset}
\alias{post_dataset}
\title{Query a data set using POST and a query json}
\usage{
post_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  filter_items = NULL,
  json_query = NULL,
  dataset_version = NULL,
  ees_environment = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 10000,
  parse = TRUE,
  debug = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-csv", "get-data" or "post-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographies}{String, vector or data frame containing the geographic levels and
locations to be queried.}

\item{filter_items}{Filter items required as a string or vector of strings}

\item{json_query}{Optional path to a json file containing the query parameters}

\item{dataset_version}{Version of data set to be connected to}

\item{ees_environment}{EES ees_environment to connect to: "dev", "test", "preprod" or "prod"}

\item{api_version}{EES API version}

\item{page}{Page number of query results to return}

\item{page_size}{Number of results to return in a single query}

\item{parse}{Logical flag to activate parsing of the results. Default: TRUE}

\item{debug}{Run POST query in debug mode. Logical, default = FALSE}

\item{verbose}{Run with additional contextual messaging. Logical, default = FALSE}
}
\value{
Data frame containing query results of an API data set
}
\description{
This function provides a method for generating and sending a json based data query to the
EES API. As a minimum, it requires the dataset_id flag and either the indicators flag or
a json file containing a query to be provided.
}
\examples{
post_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)

# Run post_dataset() to select rows containing either of two geographic locations and a single
# filter item.
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = example_id("time_period", group = "attendance"),
  geographies = example_id("location_code", group = "attendance"),
  filter_items = example_id("filter_item", group = "attendance"),
  page = 1,
  page_size = 32
)

# Run post_dataset() using set parameters giving a combination of filter options
example_id("filter_items_short", group = "attendance")
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = example_id("time_period", group = "attendance"),
  geographies = example_id("location_code", group = "attendance"),
  filter_items = example_id("filter_items_short", group = "attendance"),
)
}
