% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset.R
\name{query_dataset}
\alias{query_dataset}
\title{Query a data set}
\usage{
query_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  json_query = NULL,
  method = "POST",
  dataset_version = NULL,
  api_version = NULL,
  page_size = 1000,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "post-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{locations}{Location code required as a string or vector of strings}

\item{filter_items}{Filter items required as a string or vector of strings}

\item{json_query}{Optional path to a json file containing the query parameters}

\item{method}{An API query method. Needs to be "GET"}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame containing query results
}
\description{
Create and send a query to the EES API. Queries can be constructed by including the
codes to the relevant flags to filter on time period, geographic level, location,
and data set specific filters. If none of the above are set in the function call,
then the entire data set will be retrieved. The data set id and specific indicators
of interest must be supplied explictly using the dataset_id and indicators params.
}
\examples{
 query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator"),
  time_periods = "2024|W23",
  geographic_levels = c("NAT", "REG"),
  filter_items = c("pmRSo", "7SdXo")
)

query_dataset(
  example_id(),
  method = "GET",
  geographic_levels = "NAT",
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator")
)

query_dataset(
  example_id(),
  method = "GET",
  indicators = example_id("indicator"),
  page = 1,
  page_size = 10
)

}
