% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset.R
\name{query_dataset}
\alias{query_dataset}
\title{Query a data set}
\usage{
query_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  method = "GET",
  dataset_version = NULL,
  api_version = NULL,
  page_size = 1000,
  page = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{ID of data set to be connected to. This is required if the endpoint is one
of "get-summary", "get-meta", "get-data" or "post-data"}

\item{indicators}{Indicators required as a string or vector of strings (required)}

\item{time_periods}{Time periods required as a string ("period|code") or vector of strings}

\item{geographic_levels}{Geographic levels required as a string or vector of strings}

\item{locations}{Location code required as a string or vector of strings}

\item{filter_items}{Filter items required as a string or vector of strings}

\item{method}{An API query method. Needs to be "GET"}

\item{dataset_version}{Version of data set to be connected to}

\item{api_version}{EES API version}

\item{page_size}{Number of results to return in a single query}

\item{page}{Page number of query results to return}

\item{verbose}{Add extra contextual information whilst running}
}
\value{
Data frame containing query results
}
\description{
A short description...
}
\examples{
query_dataset(example_id(), indicators = example_id("indicator"), page = 1, page_size=10)
}
