% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset.R
\name{query_dataset_get}
\alias{query_dataset_get}
\title{Query a data set using GET and a query URL}
\usage{
query_dataset_get(dataset_id, dataset_version = NULL, api_version = NULL)
}
\arguments{
\item{api_version}{}
}
\description{
Query a data set using GET and a query URL
}
